// src/core/types/token.ts
export type Token = {
  address: string
  chainId: number
  name: string
  symbol: string
  image: string
  description: string
  decimals: number
  currentHolderCount: string
  transferCount: string
  mintCount: string
  burnCount: string
  totalSupply: string
  socials: string
  creationTimestamp: number
  priceInDecimal: string
  owner: string
  ownerName: string
  ownerImage: string
  ownerSocial: string
}